refactor tests and restructure service packages; consolidate mock setups and improve test readability
