+++ pom.xml
@@ -6,7 +6,7 @@
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>3.4.2</version>
-		<relativePath/> <!-- lookup parent from repository -->
+		<relativePath/>
 	</parent>
 	<groupId>dika</groupId>
 	<artifactId>spring.security</artifactId>
@@ -121,14 +121,12 @@
 				<artifactId>jacoco-maven-plugin</artifactId>
 				<version>0.8.8</version>
 				<executions>
-					<!-- prepare-agent встраивает Java-агент Jacoco -->
 					<execution>
 						<id>prepare-agent</id>
 						<goals>
 							<goal>prepare-agent</goal>
 						</goals>
 					</execution>
-					<!-- report генерирует HTML-отчёт (фаза verify или prepare-package) -->
 					<execution>
 						<id>report</id>
 						<phase>verify</phase>
@@ -139,13 +137,11 @@
 				</executions>
 			</plugin>
 
-			<!-- Плагин maven-surefire-plugin, который запускает тесты -->
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-surefire-plugin</artifactId>
 				<version>2.22.2</version>
 				<configuration>
-					<!-- Важная настройка: argLine позволяет Jacoco дописать -javaagent -->
 					<argLine>${argLine}</argLine>
 				</configuration>
 			</plugin>
