+++ src/test/java/dika/spring/security/mapper/UserMapperTest.java
@@ -1,5 +1,6 @@
 package dika.spring.security.mapper;
 
+import dika.spring.security.UserMapper;
 import dika.spring.security.dto.LinksEntityDto;
 import dika.spring.security.dto.reqest.UserRequestDto;
 import dika.spring.security.dto.response.UserResponseDto;
@@ -101,87 +102,7 @@ public class UserMapperTest {
         assertEquals(linksEntityDto.getTgRef(), result.getLinksEntityDTO().getTgRef());
         assertEquals(linksEntityDto.getVkRef(), result.getLinksEntityDTO().getVkRef());
     }
-
-    @Test
-    public void toDTOTestSomeLinksNull() {
-        UUID externalId = UUID.randomUUID();
-        User user = new User();
-        user.setExternalId(externalId);
-        user.setId(1L);
-        user.setUsername("Anacondaz");
-        user.setPassword("123");
-        user.setRole(List.of(Roles.USER));
-
-        LinksEntity linksEntity = new LinksEntity();
-        linksEntity.setVkRef(null);
-        linksEntity.setInstRef("qwerty");
-        linksEntity.setTgRef(null);
-
-        user.setLinksEntity(linksEntity);
-
-        LinksEntityDto linksEntityDto = new LinksEntityDto();
-        linksEntityDto.setVkRef(null);
-        linksEntityDto.setInstRef("qwerty");
-        linksEntityDto.setTgRef(null);
-
-        UserResponseDto userResponseDto = new UserResponseDto();
-        userResponseDto.setLinksEntityDTO(linksEntityDto);
-        userResponseDto.setPassword("123");
-        userResponseDto.setUsername("Anacondaz");
-        userResponseDto.setExternalId(externalId);
-        userResponseDto.setRole(List.of(Roles.USER));
-
-        UserResponseDto result = mapper.toDTO(user);
-        assertNotNull(result);
-        assertEquals(userResponseDto.getPassword(), result.getPassword());
-        assertEquals(userResponseDto.getRole(), result.getRole());
-        assertEquals(userResponseDto.getUsername(), result.getUsername());
-        assertEquals(userResponseDto.getExternalId(), result.getExternalId());
-
-        assertEquals(linksEntityDto.getInstRef(), result.getLinksEntityDTO().getInstRef());
-        assertEquals(linksEntityDto.getTgRef(), result.getLinksEntityDTO().getTgRef());
-        assertEquals(linksEntityDto.getVkRef(), result.getLinksEntityDTO().getVkRef());
-    }
-
-    @Test
-    public void fromDTOTestSomeLinksNull() {
-        UUID externalId = UUID.randomUUID();
-        User user = new User();
-        user.setExternalId(externalId);
-        user.setId(1L);
-        user.setUsername("Anacondaz");
-        user.setPassword("123");
-        user.setRole(List.of(Roles.USER));
-
-        LinksEntity linksEntity = new LinksEntity();
-        linksEntity.setVkRef(null);
-        linksEntity.setInstRef("qwerty");
-        linksEntity.setTgRef(null);
-
-        user.setLinksEntity(linksEntity);
-
-        LinksEntityDto linksEntityDto = new LinksEntityDto();
-        linksEntityDto.setVkRef(null);
-        linksEntityDto.setInstRef("qwerty");
-        linksEntityDto.setTgRef(null);
-
-        UserRequestDto userRequestDto = new UserRequestDto();
-        userRequestDto.setLinksEntityDTO(linksEntityDto);
-        userRequestDto.setPassword("123");
-        userRequestDto.setUsername("Anacondaz");
-        userRequestDto.setRole(List.of(Roles.USER));
-
-        User result = mapper.fromDTO(userRequestDto);
-        assertNotNull(result);
-        assertEquals(userRequestDto.getPassword(), result.getPassword());
-        assertEquals(userRequestDto.getRole(), result.getRole());
-        assertEquals(userRequestDto.getUsername(), result.getUsername());
-
-        assertEquals(linksEntityDto.getInstRef(), result.getLinksEntity().getInstRef());
-        assertEquals(linksEntityDto.getTgRef(), result.getLinksEntity().getTgRef());
-        assertEquals(linksEntityDto.getVkRef(), result.getLinksEntity().getVkRef());
-    }
-
+    
     @Test
     public void fromDTOTest() {
         UUID externalId = UUID.randomUUID();
@@ -292,44 +213,6 @@ public class UserMapperTest {
         linksEntityDto.setTgRef("qwerty");
 
         LinksEntity result = mapper.fromDTO(linksEntityDto);
-
-        assertEquals(linksEntityDto.getInstRef(), result.getInstRef());
-        assertEquals(linksEntityDto.getTgRef(), result.getTgRef());
-        assertEquals(linksEntityDto.getVkRef(), result.getVkRef());
-    }
-
-    @Test
-    public void fromDtoLSomeLinksNull() {
-        LinksEntity linksEntity = new LinksEntity();
-        linksEntity.setVkRef(null);
-        linksEntity.setInstRef("qwerty");
-        linksEntity.setTgRef(null);
-
-        LinksEntityDto linksEntityDto = new LinksEntityDto();
-        linksEntityDto.setVkRef(null);
-        linksEntityDto.setInstRef("qwerty");
-        linksEntityDto.setTgRef(null);
-
-        LinksEntity result = mapper.fromDTO(linksEntityDto);
-
-        assertEquals(linksEntityDto.getInstRef(), result.getInstRef());
-        assertEquals(linksEntityDto.getTgRef(), result.getTgRef());
-        assertEquals(linksEntityDto.getVkRef(), result.getVkRef());
-    }
-
-    @Test
-    public void toDtoLSomeLinksNull() {
-        LinksEntity linksEntity = new LinksEntity();
-        linksEntity.setVkRef(null);
-        linksEntity.setInstRef("qwerty");
-        linksEntity.setTgRef(null);
-
-        LinksEntityDto linksEntityDto = new LinksEntityDto();
-        linksEntityDto.setVkRef(null);
-        linksEntityDto.setInstRef("qwerty");
-        linksEntityDto.setTgRef(null);
-
-        LinksEntityDto result = mapper.toDTO(linksEntity);
 
         assertEquals(linksEntityDto.getInstRef(), result.getInstRef());
         assertEquals(linksEntityDto.getTgRef(), result.getTgRef());
